topic_cluster_app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
scikit-learn
numpy
gensim
hdbscan
umap-learn
sentence-transformers
transformers
spacy
nltk
textblob


# pip install -r requirements.txt